Version Packages (#5)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.


# Releases
## @toomuchdesign/json-schema-fns@0.1.0

### Minor Changes

-   52d8fa0: Initial release

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>